Break parts of input.c into new chunk.c file
